tree-sitter: split tree-sitter lua interface from demo code
